feat-redis - 수정본
